backend/requirements.txt
Django==5.1.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
Pillow==11.1.0
gunicorn==23.0.0
python-dotenv==1.0.1
dj-database-url==2.2.0
whitenoise==6.7.0
